08-08-2018 added entities in map service

+++ account-service/src/main/java/pl/mycar/accountservice/persistence/repository/AccountRepository.java
@@ -7,7 +7,7 @@ import pl.mycar.accountservice.persistence.entity.AccountEntity;
 import java.util.Optional;
 
 @Repository
-public interface AccountRepository extends JpaRepository<AccountEntity, Integer> {
+public interface AccountRepository extends JpaRepository<AccountEntity, Long> {
   Optional<AccountEntity> findByUsername(String username);
 
   Boolean existsByUsername(String username);
